give Patch a origin_sid
Somewhere down at the resource-level, we want to
send res-tainted events to the notify channel.
These events reference a resource in their data and
use the origin_sid of the patch as the Event's sid.
The Notify handler will use this sid and not forward
the event to the originating session.