refactor Context object
The function and design of the context object has been overhauled.
Every event holds a Context object which defines the contact which
the event should be handled in.
The context contains pointers to store, event-braodcast and other
facilities. Every receiver of a event can rely on the fact the the
context will contain all properties it needs to operate