refactor event-handling/-passing
Before there was a lose definition of the flow of Event instances. A
Conn would usually just pipe them down to the sessionhub-channel.
Broadcasting was another channel and seperate processor that piped into
the sessionhub.

This commit introduces a EventHandler interface. All major passes of an
event are implementations of EventHandler. A Server itself is the
magical, first handler for new events from some source (tcp, websocket)

The server passes the event onto the token-consumer eventhandler and
then down to the sessionhub's router.  Sessions are also event handlers
and the next pass for an event after the sessionhub.

The Upstream client is also an eventhandler, where any pass inbetween
can chose to push events.

Broadcasting (formerly notify) is supported via a Broadcast(..) event
handler of sessionhub.