add event remarks, notify token consumation errors
event messgaes now contain a new property 'remark' which is an object
containing 'lvl', 'msg' and 'data' fields that describe a remark.
A remark always only references an event with the same Tag, e.g. if the
client sends a `token-consume` event with event.tag = `abcfoo` and the
token is invalid, the response `token-consume` event will contain the
same tag (`abcfoo`) and a remark (explaining why the token cannot be
consumed).
There is no guarantee that the event which contains the remark will
actually reach the client or will be processed by it.
Sending remarks is a fire and forget action and no state is kept
whatsoever