add remarks to failed `res-sync` requests
If a fatal error occurs during a sync event (e.g. SV mismatch), the
server will abort the current sync cycle and respond to the res-sync
event with an empty changes array and a non-empty remark object.

If the error is not fatal, but recoverable the server also responds with
an empty changes list and a remark object, but the level is 'error'
instead of 'fatal'