add session lifetime and status checks
This commit adds timeouts for sessions. If a sessions reaches the end of
its lifetime, it is considered invalid. If any action tries to access it
directly (e.g. sends an event to it) the response will contain a 'fatal'
remark with the slug 'sesion-expired'.

Also a session can be marked as 'terminated' which means that it cannot
be used for any sync events, but can still be used to re-create a new
login-token via the /tokens/login endpoint (with only the `sid` as
payload).
This is useful in case the client wants to reset the session. A post to
/tokens/login will implictly mark the old sid as 'terminated'.
The old sessionrunner will not immediately shutdown. As long as it is
alive it will still accept and process any sync event it receives.

After an active runner of a 'terminated' session has shut down, no new
runner can be started for that session-id.